[example app] updated yarn.lock so latest Theia used
The example application that comes with this repo was still
using an earlier Theia version, because of yarn.lock. I
have manually removed the @theia entries and re-ran yarn:
now the latest version of Theia packages are picked-up.

As well, with this change, we are now avoiding the example
application using a moderate severity security vulnerability,
present in dependency "fstream < 1.0.12".

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>